changed create, fill and metric history
added template title to create metric. it is now fully functional and works with the history. attempting to add choose a template to the fill metric side. as of right now its not working. saved forms (aka metric history) is functional. none of this includes brandon's updated code from last night